pyproject.toml
[project]
name = "zotero.notion-ghstars"
version = "0.1.0"
description = "Sync GitHub repository links and star counts into a Notion paper database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "notion-client>=2.7.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "anyio>=4.12.0",
    "pytest>=8.4.1",
]

[tool.uv]
package = false

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.pytest.ini_options]
pythonpath = ["."]
